Public assault prediction using Spatio-Temporal Net algorithm
In this project, KTH video dataset is used for designing the system. 
Preprocessing technique is used to extract the certain frames in dataset. 
In feature extraction process, Pixel and Optical flow feature extraction techniques are used to extract the features. 
Data visualization method is used for visualize the feature extraction. 
The deep learning algorithm such as Spatio-Temporal Net is then used to determine and classify the activities of a human.  
We can provide this efficient model as an application to road surveillance as such a camera module fixed in the road to perform constant surveillance. 
The camera on recognizing abnormal detection from the humans such as fights, etc., an alert notification is sent to the police. 
A mobile application is developed using react native which will be held by the police to which, the camera on recognizing abnormality in the humans an alert notification is sent and live streaming is enabled. 
Thus, this project successfully provides a Human activity recognition model incorporating AI to the cameras which can be used in real time applications such as a solution to prevent and provide evidence of an abnormal detection in road.